+++ RELEASE-NOTES.md
@@ -11,7 +11,7 @@
 	* `Statement::setClaim` and `Statement::getClaim` have been removed
 	* Removed `ClaimList`
 	* Removed `ClaimListAccess`
-	* Removed `addClaim`, `hasClaims` and `newClaim` from all entity classes
+	* Removed `addClaim`, `getClaims`, `hasClaims`, `newClaim` and `setClaims` from all entity classes
 * `Claims::addClaim` no longer supports setting an index
 * Removed `Claims::getBestClaims` (you can use `StatementList::getBestStatements` instead)
 * Removed `Claims::getByRank` and `Claims::getByRanks` (you can use `StatementList::getWithRank` instead)